Extract PersonContentGridView.FeedLoaderType to new file
